requirements.txt
keyboard==0.13.5
matplotlib==3.7.1
numpy==1.24.3
opencv_python==4.7.0.72
pandas==2.0.1
Pillow==9.5.0
Pillow==9.5.0
scikit_learn==1.2.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
